@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
+    "build": "vue-cli-service build --inline-vue",
     "lint": "vue-cli-service lint",
     "test:e2e": "vue-cli-service test:e2e",
     "test:unit": "vue-cli-service test:unit",
@@ -12,52 +12,42 @@
     "semantic-release": "semantic-release"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.12",
-    "@fortawesome/free-solid-svg-icons": "^5.6.3",
-    "@fortawesome/vue-fontawesome": "^0.1.5",
-    "aes256": "^1.0.4",
-    "bech32": "^1.1.3",
-    "bootstrap": "^4.3.1",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/vue-fontawesome": "^3.0.1",
+    "@popperjs/core": "^2.11.5",
+    "@vue/compat": "^3.2.37",
+    "aes256": "^1.1.0",
+    "bech32": "^2.0.0",
+    "bootstrap": "^5.1.3",
+    "browserify-fs": "^1.0.0",
+    "buffer": "^6.0.3",
     "crypto": "^1.0.1",
+    "crypto-browserify": "^3.12.0",
     "hack-font": "^3.3.0",
     "html2pdf": "0.0.11",
-    "html2pdf.js": "^0.9.1",
-    "jquery": "^3.5.0",
-    "jspdf": "^2.3.1",
-    "popper.js": "^1.14.6",
-    "print-js": "^1.0.54",
-    "qrllib": "^1.0.4",
+    "html2pdf.js": "^0.10.1",
+    "jquery": "^3.6.0",
+    "jspdf": "^2.5.1",
+    "path-browserify": "^1.0.1",
+    "print-js": "^1.6.0",
+    "qrllib": "^1.2.4",
     "roboto-fontface": "^0.10.0",
-    "vue": "^2.5.21",
-    "vue-router": "^3.0.1",
-    "vuex": "^3.0.1"
+    "stream-browserify": "^3.0.0",
+    "util": "^0.12.4",
+    "vue": "^3.2.37",
+    "vue-router": "^4.0.16",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@semantic-release/changelog": "^5.0.1",
-    "@semantic-release/git": "^9.0.0",
-    "@semantic-release/github": "^7.2.3",
-    "@semantic-release/npm": "^7.1.3",
-    "@vue/cli-plugin-babel": "^3.3.0",
-    "@vue/cli-plugin-e2e-cypress": "^3.3.0",
-    "@vue/cli-plugin-eslint": "^3.3.0",
-    "@vue/cli-plugin-unit-jest": "^3.3.0",
-    "@vue/cli-service": "3.3.0",
-    "@vue/eslint-config-airbnb": "^4.0.0",
-    "@vue/test-utils": "^1.0.0-beta.20",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^23.6.0",
-    "common-config-webpack-plugin": "^1.2.1",
-    "css-loader": "^2.1.0",
-    "eslint": "^5.8.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "file-loader": "^3.0.1",
-    "html-webpack-plugin": "^3.2.0",
-    "semantic-release": "^17.4.3",
-    "style-loader": "^0.23.1",
-    "url-loader": "^1.1.2",
-    "vue-template-compiler": "^2.5.21",
-    "webpack": "^4.28.4",
-    "webpack-cli": "^3.2.1"
+    "@semantic-release/changelog": "^6.0.1",
+    "@semantic-release/git": "^10.0.1",
+    "@semantic-release/github": "^8.0.4",
+    "@semantic-release/npm": "^9.0.1",
+    "@vue/cli-service": "^5.0.6",
+    "@vue/compiler-sfc": "^3.2.37",
+    "html-webpack-inline-source-plugin": "^0.0.10",
+    "html-webpack-plugin": "^5.5.0",
+    "semantic-release": "^19.0.3"
   }
 }
