@@ -1,40 +1,42 @@
-import Vue from 'vue';
-import Router from 'vue-router';
-import Home from './views/Home.vue';
+import Vue from "vue";
+import { createRouter, createWebHistory } from "vue-router";
+import Home from "./views/Home.vue";
 
-Vue.use(Router);
+// Vue.use(Router);
 
-export default new Router({
-  routes: [{
-    path: '/',
-    name: 'home',
-    component: Home,
-    meta: {
-      title: 'QRL Offline Wallet Generator'
+export default new createRouter({
+  history: createWebHistory(),
+  routes: [
+    {
+      path: "/",
+      name: "home",
+      component: Home,
+      meta: {
+        title: "QRL Offline Wallet Generator",
+      },
     },
-  },
-  {
-    path: '/about',
-    name: 'about',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ './views/About.vue'),
-    meta: {
-      title: 'QRL Offline Wallet Generator'
+    {
+      path: "/about",
+      name: "about",
+      // route level code-splitting
+      // this generates a separate chunk (about.[hash].js) for this route
+      // which is lazy-loaded when the route is visited.
+      component: () => import(/* webpackChunkName: "about" */ "./views/About.vue"),
+      meta: {
+        title: "QRL Offline Wallet Generator",
+      },
     },
-  },
-  {
-    path: '/docs',
-    name: 'docs',
-    // route level code-splitting
-    // this generates a separate chunk (docs.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "docs" */ './views/Docs.vue'),
-    meta: {
-      title: 'QRL Offline Wallet Generator'
+    {
+      path: "/docs",
+      name: "docs",
+      // route level code-splitting
+      // this generates a separate chunk (docs.[hash].js) for this route
+      // which is lazy-loaded when the route is visited.
+      component: () => import(/* webpackChunkName: "docs" */ "./views/Docs.vue"),
+      meta: {
+        title: "QRL Offline Wallet Generator",
+      },
     },
-  },
   ],
-  linkActiveClass: 'active',
+  linkActiveClass: "active",
 });
