@@ -1,7 +1,7 @@
-import Vue from 'vue';
+// import Vue from 'vue';
 import Vuex from 'vuex';
 
-Vue.use(Vuex);
+// Vue.use(Vuex);
 
 export default new Vuex.Store({
   state: {
